scraper/requirements.txt
supabase
requests
beautifulsoup4
pandas
playwright
playwright-stealth
google-generativeai
